Move `cylc remove` method from task pool to scheduler
This will allow it to call the method to kill submitted/running tasks
